Create README.md
# iOS Movie Trending List App

This is a simple iOS application demonstrating the use of the **Model-View-ViewModel (MVVM)** pattern in Swift. 
The app fetches and displays a list of trending movies. 
The app is designed as an educational project based on the **Swift Academy YouTube** channel's tutorial series. 
This repository showcases core iOS development concepts using the MVVM architecture.